Reorganizations
Updated to a newer version of pgx, upgraded the setupper to a modified version of the kessler one, organized the assets file system (all in lowercase for compatibility), and made the game support mac a bit better with resolution.

Still need to reorganize buttons and make a working fullscreen toggle in the options screen.